requirements.txt
# Backend Dependencies
# fastapi==0.104.1
# uvicorn==0.24.0
# pydantic==2.4.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9

# Frontend Dependencies
streamlit==1.51.0
requests==2.31.0

# AI Dependencies
# scikit-learn==1.3.2
numpy==1.26.1
pandas==2.1.2

# Testing Dependencies
# pytest==7.4.3
# httpx==0.25.1

# Development Dependencies
# black==23.10.1
# flake8==6.1.0
# isort==5.12.0